third_party/WebKit/LayoutTests/crypto/aes-gcm-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests bad algorithm inputs for AES-GCM");

jsTestIsAsync = true;

var keyData = hexStringToUint8Array("2b7e151628aed2a6abf7158809cf4f3c");
var data = asciiToUint8Array("hello");
var key = null;

Promise.resolve(null).then(function(result) {
    var usages = ['encrypt', 'decrypt'];
    var extractable = false;
    var algorithm = {name: 'aes-gcm'};

    return crypto.subtle.importKey('raw', keyData, algorithm, extractable, usages);
}).then(function(result) {
    key = result;

    return crypto.subtle.encrypt({name: 'AES-gcm'}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AES-gcm', iv: 3}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AES-gcm', iv: 'foo'}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AeS-gcm', iv: new Uint8Array(16), additionalData: '5'}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: 'foo'}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: -1}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: 8000}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
